+++ ext/fs/Cargo.toml
@@ -14,15 +14,13 @@ description = "Ops for interacting with the file system"
 path = "lib.rs"
 
 [features]
-sync_fs = ["deno_config/sync"]
+sync_fs = []
 
 [dependencies]
 async-trait.workspace = true
 base32.workspace = true
-deno_config = { workspace = true, default-features = false }
 deno_core.workspace = true
 deno_io.workspace = true
-deno_package_json.workspace = true
 deno_permissions.workspace = true
 filetime.workspace = true
 libc.workspace = true
